tiler_python/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
openslide-python==1.3.0
Pillow==10.1.0
boto3==1.34.0
cachetools>=5.5.2
